refactor(registration-react): extract datasetRegistrationConnector
